Make WorkerInfo immutable to support in-place restarts
位于同一 worker_id 空间的 server 可以通过具有相同的 `BLLM_KVTRANS_PORT_BASE` env 值来实现对于一个特定的 instid/workerid, 其 WorkerInfo 总是不可变的, 进而支持原地重启, 以及后续的 refine txstub thread model.

可变的信息可以通过建联协商获取, 就如当前的 rdma layer mr handle.
Link: https://code.alibaba-inc.com/PAI-LLM/blade-kvt/codereview/23782687
* Make WorkerInfo immutable to support in-place restarts

+++ kvtransfer/include/common.h
@@ -95,8 +95,8 @@ struct WorkerInfo {
   WorkerId worker_id;
   uint32_t tp_size;
   uint32_t worker_tp_rank;
-  uint32_t block_size;
-  uint32_t token_size;
+  size_t block_size;
+  size_t token_size;
   uint32_t layer_num_blocks{1};
   uint32_t num_layers{1};
   uint8_t transfer_protocols{0};
